+++ src/components/mod.rs
@@ -1,5 +1,3 @@
 pub mod document;
 pub mod item;
 pub mod container;
-pub mod utilities;
-mod def_constants;
