+++ src/utilities/def_constants.rs
